test(utils): split tests/test_utils.py across new module test files
Mirror the production split that happened in the preceding commits.
The ten tests in the old tests/test_utils.py move into:

  tests/test_fileops.py (new) - 7 tests:
    TestEnsureDirExists::{test_create,test_create_exists,
                         test_create_permission_denied,test_on_create},
    test_chdir, test_atomic_write, test_timestamp.

  tests/test_term.py (new) - 1 test:
    test_filter_ansi.

  tests/test_misc.py (existing) - 1 test appended:
    test_sutparse.

The redundant test_colors entry is dropped: tests/test_color.py
already exercises the four ANSI helpers under
test_green_emits_ansi_when_enabled / test_green_returns_plain_when_disabled
with identical assertions across all four colours plus both colour
modes, so the test_utils.py copy was strictly subsumed.

Net pytest collected count: 691 -> 690 (-1, the redundant test_colors).

Seventh commit of the Phase 5b / C8 split (PLAN.md item 11).
